chore: refactored src/utils/fetchMetrics
separated concerns in different files by functions and types.
NOTE: this change does not affect the codebase as everything that was exported by fetchMetrics.ts now is exported by src/utils/fetchMetrics/index.ts
